New wand and potion for Imoen
Also attempted to put house key in wall safe, but not found in inventory when party moved to AR0602.  The bag of holding give to Player1 is empty.  Not sure how that would have worked anyway, seeing as it was a key inside of a keyring.  This needs to be fixed